docs: refactor eleventy config
