+++ .eleventy.cjs
@@ -17,7 +17,6 @@ const markdownItAnchor = require('markdown-it-anchor');
 const pluginToc = require('@patternfly/pfe-tools/11ty/plugins/table-of-contents.cjs');
 const sassPlugin = require('eleventy-plugin-dart-sass');
 
-const dedent = require('./dedent.cjs');
 const path = require('path');
 
 const markdownLib = markdownIt({
@@ -93,6 +92,35 @@ module.exports = function(eleventyConfig) {
       }
     },
   });
+
+  function dedent(callSite, ...args) {
+    function format(str) {
+      let size = -1;
+
+      return str.replace(/\n(\s+)/g, (_, m1) => {
+        if (size < 0) {
+          size = m1.replace(/\t/g, '    ').length;
+        }
+
+        return `\n${m1.slice(Math.min(m1.length, size))}`;
+      });
+    }
+
+    if (typeof callSite === 'string') {
+      return format(callSite);
+    }
+
+    if (typeof callSite === 'function') {
+      return (...args) => format(callSite(...args));
+    }
+
+    const output = callSite
+      .slice(0, args.length + 1)
+      .map((text, i) => `${(i === 0 ? '' : args[i - 1])}${text}`)
+      .join('');
+
+    return format(output);
+  }
 
   /**
    * Section macro
